pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow==3.1.8",
]
